+++ src/config/mod.rs
@@ -12,10 +12,15 @@ pub struct Config {
     providers: Vec<Provider>,
     active_provider: Option<usize>,
     #[serde(default)]
+    active_search_provider: Option<usize>,
+    #[serde(default)]
     pub use_streaming: bool,
     /// **provider_name, model_id, display_name**
+    #[serde(default, rename = "available_models")]
+    available_completion_models: Vec<(ProviderName, String, String)>,
+    /// **provider_name, model_id, display_name**
     #[serde(default)]
-    available_models: Vec<(ProviderName, String, String)>,
+    available_search_models: Vec<(ProviderName, String, String)>,
     #[serde(default)]
     mcp_clients: Vec<McpClient>,
 }
@@ -51,32 +56,44 @@ impl Config {
         self.providers.iter().any(|p| p.name() == provider_name)
     }
 
-    pub fn active_model(&self) -> Option<(String, String)> {
-        if let Some(model) = self.active_provider().map(|p| p.model()) {
-            let openai_models: &[(&str, &str)] = llm_models::OPENAI_MODELS;
-            let anthropic_models: &[(&str, &str)] = llm_models::ANTHROPIC_MODELS;
+    /// Get `(model_id, display_name)` of the active model
+    pub fn active_model(&self, search: bool) -> Option<(String, String)> {
+        let provider = match search {
+            true => self.active_search_provider()?,
+            false => self.active_provider()?,
+        };
 
-            // Find the model in the list of available models
-            let model = match self.active_provider()? {
-                Provider::OpenAI(_) => openai_models.iter().find(|(id, _)| id == &model),
-                Provider::Anthropic(_) => anthropic_models.iter().find(|(id, _)| id == &model),
-            };
+        let active_model = match search {
+            true => provider.search_model()?,
+            false => provider.completion_model(),
+        };
 
-            model.map(|(id, name)| (id.to_string(), name.to_string()))
-        } else {
-            None
-        }
+        let available_models: &[(&str, &str)] = match (provider, search) {
+            (Provider::OpenAI(_), true) => llm_models::OPENAI_SEARCH_MODELS,
+            (Provider::OpenAI(_), false) => llm_models::OPENAI_COMPLETION_MODELS,
+            (Provider::Anthropic(_), true) => llm_models::ANTHROPIC_SEARCH_MODELS,
+            (Provider::Anthropic(_), false) => llm_models::ANTHROPIC_COMPLETION_MODELS,
+        };
+
+        available_models
+            .iter()
+            .find(|(id, _)| id == &active_model)
+            .map(|(id, display_name)| (id.to_string(), display_name.to_string()))
     }
 
     pub fn active_provider(&self) -> Option<&Provider> {
         self.providers.get(self.active_provider?)
     }
 
+    pub fn active_search_provider(&self) -> Option<&Provider> {
+        self.providers.get(self.active_search_provider?)
+    }
+
     pub fn find_provider(&self, provider_name: &ProviderName) -> Option<&Provider> {
         self.providers.iter().find(|p| &p.name() == provider_name)
     }
 
-    pub fn set_model(&mut self, provider_name: ProviderName, model: String) {
+    pub fn set_completion_model(&mut self, provider_name: ProviderName, model: String) {
         self.active_provider = self
             .providers
             .iter()
@@ -84,14 +101,29 @@ impl Config {
 
         if let Some(provider) = self.active_provider() {
             let mut provider = provider.clone();
-            provider.set_model(model);
+            provider.set_completion_model(model);
             self.providers[self.active_provider.unwrap()] = provider;
             self.save();
         }
     }
 
+    pub fn set_search_model(&mut self, provider_name: ProviderName, model: String) {
+        self.active_search_provider = self
+            .providers
+            .iter()
+            .position(|p| p.name() == provider_name);
+
+        if let Some(provider) = self.active_search_provider() {
+            let mut provider = provider.clone();
+            provider.set_search_model(model);
+            self.providers[self.active_search_provider.unwrap()] = provider;
+            self.save();
+        }
+    }
+
     pub async fn refresh_available_models(&mut self) {
-        let mut models: Vec<(ProviderName, String, String)> = Vec::new();
+        let mut completion_models: Vec<(ProviderName, String, String)> = Vec::new();
+        let mut search_models: Vec<(ProviderName, String, String)> = Vec::new();
 
         let mut tasks = vec![];
 
@@ -102,18 +134,27 @@ impl Config {
 
         let results = futures::future::join_all(tasks).await;
 
-        for (i, result) in results.into_iter().enumerate() {
+        for (i, (cm, sm)) in results.into_iter().enumerate() {
             let provider = &self.providers[i];
-            for (id, display_name) in result {
-                models.push((provider.name(), id, display_name));
+            for (id, display_name) in cm {
+                completion_models.push((provider.name(), id, display_name));
+            }
+
+            for (id, display_name) in sm {
+                search_models.push((provider.name(), id, display_name));
             }
         }
 
-        self.available_models = models;
+        self.available_completion_models = completion_models;
+        self.available_search_models = search_models;
     }
 
-    pub fn get_available_models(&self) -> &Vec<(ProviderName, String, String)> {
-        &self.available_models
+    pub fn available_completion_models(&self) -> &Vec<(ProviderName, String, String)> {
+        &self.available_completion_models
+    }
+
+    pub fn available_search_models(&self) -> &Vec<(ProviderName, String, String)> {
+        &self.available_search_models
     }
 
     pub fn remove_provider(&mut self, provider_name: ProviderName) {
@@ -123,7 +164,9 @@ impl Config {
             .position(|p| p.name() == provider_name);
 
         self.providers.retain(|p| p.name() != provider_name);
-        self.available_models
+        self.available_completion_models
+            .retain(|(p, _, _)| p != &provider_name);
+        self.available_search_models
             .retain(|(p, _, _)| p != &provider_name);
 
         if let Some(index) = self.active_provider {
@@ -148,10 +191,14 @@ impl Config {
             provider.set_api_key(api_key);
             self.providers[index] = provider;
         } else {
-            let provider = Provider::new(provider_name, api_key, None).await;
-            let models = provider.fetch_available_models().await;
-            for (id, display_name) in models {
-                self.available_models
+            let provider = Provider::new(provider_name, api_key).await;
+            let (completion_models, search_models) = provider.fetch_available_models().await;
+            for (id, display_name) in completion_models {
+                self.available_completion_models
+                    .push((provider_name, id, display_name));
+            }
+            for (id, display_name) in search_models {
+                self.available_search_models
                     .push((provider_name, id, display_name));
             }
             self.providers.push(provider);
@@ -186,7 +233,8 @@ impl Config {
         let mut cfg = self.clone();
 
         cfg.providers.sort_by_key(|a| a.name());
-        cfg.available_models.sort_by_key(|a| a.0);
+        cfg.available_completion_models.sort_by_key(|a| a.0);
+        cfg.available_search_models.sort_by_key(|a| a.0);
 
         for provider in cfg.providers.iter_mut() {
             match provider.encrypt() {